small changes after remarks
